BE-2: id changed to Long, exception classname changed, checkstyle added

+++ src/main/java/az/gdg/msteam/model/entity/MemberEntity.java
@@ -3,11 +3,8 @@ package az.gdg.msteam.model.entity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import javax.persistence.Table;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
+
+import javax.persistence.*;
 
 @Data
 @AllArgsConstructor
@@ -18,7 +15,7 @@ public class MemberEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String firstName;
     private String lastName;
